orchestrator/requirements.txt
PyYAML
transitions
asyncpg
SQLAlchemy[asyncio]
sqlalchemy>=2.0
asyncpg
aiokafka
